+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - When the current token has an open single-quote (`'`), fish will now escape any `'` and `\` in pasted text so that it can be used as a single token. Note that this requires either bracketed paste or use of the special `fish_clipboard_paste` function (bound to \cv by default).
 - Fish is now more forgiving of missing or invalid $TERM values (#3850).
 - The `string` command now supports a `repeat` subcommand with the obvious behavior (#3864).
-- The `string match` command now supports a `--filter` flag to emit the entire string partially matched by a pattern (#3957).
+- The `string match` command now supports a `--entire` flag to emit the entire string partially matched by a pattern (#3957).
 - The `string replace` command now supports a `--filter` flag to limit output to strings which underwent a replacement (#3348).
 - The `functions --details --verbose` output now includes the function description (#597).
 - Completions for `helm` added (#3829).